requirements.txt
fastapi
uvicorn
llama-index
llama-index-readers-docling
llama-index-embeddings-ollama
llama-index-vector-stores-postgres
llama-index-storage-docstore-postgres
llama-index-storage-docstore-postgres
llama-index-llms-ollama
python-dotenv
crewai
crewai-tools
ragas
datasets
rank_bm25
llama-index-retrievers-bm25
docling-ibm-models
sentencepiece
psycopg2-binary
openinference-instrumentation-crewai
arize-phoenix-otel
arize-phoenix
openinference-instrumentation-litellm
sentence-transformers
-e .